+++ CHANGELOG.md
@@ -6,13 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 - PeriodicExportingMetricReader will continue if collection times out 
   ([#3100](https://github.com/open-telemetry/opentelemetry-python/pull/3100))
 - Fix formatting of ConsoleMetricExporter.
   ([#3197](https://github.com/open-telemetry/opentelemetry-python/pull/3197))
-
 - Implement shutdown procedure forOTLP grpc exporters
   ([#3138](https://github.com/open-telemetry/opentelemetry-python/pull/3138))
+- Add exponential histogram
+  ([#2964](https://github.com/open-telemetry/opentelemetry-python/pull/2964))
 
 ## Version 1.16.0/0.37b0 (2023-02-17)